requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5
Django==2.0.2
django-cors-headers==2.1.0
django-extensions==1.9.9
django-rest-swagger==2.1.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyJWT==1.5.3
pytz==2017.3
requests==2.18.4
simplejson==3.14.0
six==1.11.0
typing==3.6.4
uritemplate==3.0.0
urllib3==1.22
coveralls==1.3.0
PyYAML==3.12
